requirements.txt
Flask
Flask-SQLAlchemy
flask-migrate
python-fedora
